Add new Decimal class
Previously decimal-related functionality existed in the Ratio class which was
the wrong place for it. This has been refactored to a new class dedicated to
decimal functionality.

In future it might be worth considering some traits or a common base class for
Fraction, Ratio and Decimal as they are all essentially the same thing
expressed in different ways.